add exception for wrong parameter type in request
